[Day05] Part B - Missing seat
Added solution for part B of day 05.

Also fixed a bug where last character of the row identifier (7th
character in the boarding pass) was not being considered. Took a while
to find this one! Also, learnt a lesson that array notations are
exclusive, not inclusive i.e. a[:n] means start to n-1 (and not start to
n)!